+++ src/plugins/query_enhancements/public/query_assist/components/query_assist_bar.tsx
@@ -15,7 +15,7 @@ import { useOpenSearchDashboards } from '../../../../opensearch_dashboards_react
 import { QueryAssistParameters } from '../../../common/query_assist';
 import { getStorage } from '../../services';
 import { useGenerateQuery } from '../hooks';
-import { getPersistedLog, AgentError, ProhibitedQueryError, appendQueryPrompt } from '../utils';
+import { getPersistedLog, AgentError, ProhibitedQueryError } from '../utils';
 import { QueryAssistCallOut, QueryAssistCallOutType } from './call_outs';
 import { QueryAssistInput } from './query_assist_input';
 import { QueryAssistSubmitButton } from './submit_button';
@@ -67,8 +67,7 @@ export const QueryAssistBar: React.FC<QueryAssistInputProps> = (props) => {
     previousQuestionRef.current = inputRef.current.value;
     persistedLog.add(inputRef.current.value);
     const params: QueryAssistParameters = {
-      // Only append query prompt in request payload
-      question: appendQueryPrompt(inputRef.current.value),
+      question: inputRef.current.value,
       index: selectedIndex,
       language: props.dependencies.language,
       dataSourceId: selectedDataset?.dataSource?.id,